ref: remove MUI out of the project
