+++ package.json
@@ -24,8 +24,6 @@
     "@emotion/styled": "11.11.0",
     "@headlessui/react": "1.7.17",
     "@hookform/resolvers": "3.1.1",
-    "@mui/icons-material": "5.14.1",
-    "@mui/material": "5.14.1",
     "@radix-ui/react-dropdown-menu": "2.0.6",
     "@radix-ui/react-slider": "^1.1.2",
     "@react-oauth/google": "0.11.1",
@@ -35,6 +33,7 @@
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
     "@typescript-eslint/parser": "6.2.1",
+    "clsx": "^2.1.0",
     "core-js": "^3.35.0",
     "eslint": "8.45.0",
     "eslint-config-next": "13.4.11",
