--- src/features/profile-setting/ui/payments/payments.stories.tsx
@@ -1,21 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { Payments } from './payments'
-
-import { payments, Table } from '@/shared/ui'
-
-const meta: Meta<typeof Payments> = {
-  component: Payments,
-  title: 'Components/Payments',
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-}
-
-export default meta
-type Story = StoryObj<typeof Payments>
-
-export const PaymentStory = () => {
-  return <Table items={payments} />
-}
