+++ src/shared/assets/icons/index.ts
@@ -3,3 +3,4 @@ export * from '@/shared/assets/icons/sort-asc/sort-asc'
 export * from '@/shared/assets/icons/sort-desc/sort-desc'
 export * from './banned/banned-icon'
 export * from './arrow-back-icon/arrow-back'
+export { Notification as NotificationIcon } from './notification/notification'
