--- src/shared/ui/_table/table-skeleton/table-skeleton.module.scss
@@ -1,45 +1,0 @@
-$base-color: #262424;
-$shine-color: rgb(56 56 56 / 80%);
-$animation-duration: 2s;
-$avatar-offset: 52 + 16;
-
-@mixin background-gradient {
-  background-image: linear-gradient(90deg, $base-color 0, $shine-color 40px, $base-color 80px);
-  background-size: 600px;
-}
-
-.tg {
-  background-color: var(--color-dark-500);
-  border: 1px solid var(--color-dark-500);
-}
-
-.tg td {
-  background-color: var(--color-dark-500);
-}
-
-.tg th {
-  background-color: var(--color-dark-500);
-}
-
-.line {
-  @include background-gradient;
-
-  float: left;
-
-  width: 100px;
-  height: 16px;
-
-  border-radius: 7px;
-
-  animation: shine-lines $animation-duration infinite ease-out;
-}
-
-@keyframes shine-lines {
-  0% {
-    background-position: -100px;
-  }
-  40%,
-  100% {
-    background-position: 140px;
-  }
-}
