--- src/shared/ui/_table/table.module.scss
@@ -1,57 +1,0 @@
-.table {
-  border-collapse: collapse;
-
-  width: 100%;
-
-  color: var(--color-light-100);
-
-  background-color: var(--color-dark-900);
-  border: 1px solid var(--color-dark-500);
-}
-
-.thead {
-  padding: 6px 24px;
-
-  font-size: var(--font-size-s);
-  font-weight: var(--font-weight-bold);
-  line-height: var(--line-height-m);
-
-  background-color: var(--color-dark-500) !important;
-  border: none;
-}
-
-.tbody {
-  word-break: break-word;
-}
-
-.tr {
-  border: 1px solid var(--color-dark-500);
-}
-
-.td {
-  padding: 6px 24px;
-
-  font-size: var(--font-size-s);
-  line-height: var(--line-height-m);
-
-  background-color: var(--color-dark-900);
-  border-bottom: 1px solid var(--color-dark-500);
-
-  > img {
-    width: 50px;
-  }
-}
-
-.th {
-  padding: 6px 24px;
-  text-align: left;
-}
-
-.titleAndSortIcon {
-  display: flex;
-  align-items: center;
-
-  :nth-child(1) {
-    margin-left: 3px;
-  }
-}
