+++ src/shared/ui/index.ts
@@ -10,13 +10,11 @@ export { RemoveAvatarButton } from '@/shared/ui/remove-avatar-button/remove-avat
 export { RoundCheckbox } from '@/shared/ui/round-checkbox/round-checkbox'
 export { Select } from './select/select'
 export { Switcher } from './switcher/index'
-export { Table } from './table/table'
-export { payments } from './table/constants'
 export { DropdownMenu } from '@/shared/ui/drop-down-menu/drop-down-menu'
 export { Text } from './text/text'
 export { ContentWrapper } from '@/shared/ui/content-wrapper/content-wrapper'
-export { THead, TBody, TRow, THeader, TCell, Table as NewTable } from './_table/table'
-export { TableSkeleton } from '@/shared/ui/_table/table-skeleton/table-skeleton'
+export { THead, TBody, TRow, THeader, TCell, Table as NewTable } from '@/shared/ui/table/table'
+export { TableSkeleton } from '@/shared/ui/table/table-skeleton/table-skeleton'
 export { Pagination } from './pagination/pagination'
 export { usePagination } from './pagination/use-pagination'
 export { Select as RadixSelect } from '@/shared/ui/radix-select/select'
