+++ src/shared/ui/table/table-skeleton/table-skeleton.module.scss
@@ -9,12 +9,12 @@ $avatar-offset: 52 + 16;
 }
 
 .tg {
-  background-color: var(--color-dark-900);
+  background-color: var(--color-dark-500);
   border: 1px solid var(--color-dark-500);
 }
 
 .tg td {
-  background-color: var(--color-dark-900);
+  background-color: var(--color-dark-500);
 }
 
 .tg th {
