--- src/shared/ui/table/table-skeleton/table-skeleton.stories.tsx
@@ -1,18 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-
-import { TableSkeleton as TSkeleton } from '@/shared/ui'
-
-const meta = {
-  component: TSkeleton,
-  tags: ['autodocs'],
-  title: 'UI/TableSkeleton',
-} satisfies Meta<typeof TSkeleton>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const TableSkeleton: Story = {
-  args: {
-    numRows: 10,
-  },
-}
